+++ packages/client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # fuzd-client
 
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-chain-protocol@0.9.2
+  - fuzd-scheduler@0.9.2
+  - fuzd-executor@0.9.2
+
 ## 0.9.1
 
 ### Patch Changes