+++ packages/tlock-decrypter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # fuzd-tlock-decrypter
 
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-scheduler@0.9.2
+
 ## 0.9.1
 
 ### Patch Changes