+++ platforms/nodejs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # fuzd-nodejs
 
+## 0.9.2
+
+### Patch Changes
+
+- Updated dependencies
+  - fuzd-server@0.9.2
+
 ## 0.9.1
 
 ### Patch Changes